chore(sequencer_node): move test util func
commit-id:d24a335f
